Corrigindo o bug do banco e removendo o Dao antigo.

+++ OrganizerList/app/src/main/java/br/com/organizerlist/database/converter/Converters.kt
@@ -1,17 +1,18 @@
 package br.com.organizerlist.database.converter
 
 import androidx.room.TypeConverter
+import java.math.BigDecimal
 import java.math.BigInteger
 
 class Converters {
 
     @TypeConverter
-    fun ofDouble(value: Double?) : BigInteger {
-        return value?.let { BigInteger(value.toString()) } ?: BigInteger.ZERO
+    fun ofDouble(value: Double?) : BigDecimal {
+        return value?.let { BigDecimal(value.toString()) } ?: BigDecimal.ZERO
     }
 
     @TypeConverter
-    fun bigDecimalToDouble(value: BigInteger?) : Double? {
+    fun bigDecimalToDouble(value: BigDecimal?) : Double? {
         return value?.let { value.toDouble() }
     }
 }
