+++ OrganizerList/app/src/main/java/br/com/organizerlist/extensions/BigIntegerExtensions.kt
@@ -1,10 +1,11 @@
 package br.com.organizerlist.extensions
 
+import java.math.BigDecimal
 import java.math.BigInteger
 import java.text.NumberFormat
 import java.util.*
 
-fun BigInteger.turnString(): String {
+fun BigDecimal.turnString(): String {
     val formatador: NumberFormat = NumberFormat.getNumberInstance()
     return formatador.format(this)
 }
