+++ OrganizerList/app/src/main/java/br/com/organizerlist/model/Product.kt
@@ -4,6 +4,7 @@ import android.os.Parcelable
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 import kotlinx.parcelize.Parcelize
+import java.math.BigDecimal
 import java.math.BigInteger
 
 @Entity
@@ -11,5 +12,5 @@ import java.math.BigInteger
 data class Product(
     @PrimaryKey(autoGenerate = true) val id: Long = 0L,
     val name: String,
-    val qty: BigInteger
+    val qty: BigDecimal
 ) : Parcelable
